+++ napari_live_recording/Cameras/CameraOpenCV.py
@@ -1,4 +1,4 @@
-from .ICamera import ICamera
+from .ICamera import CameraROI, ICamera
 from platform import system
 import cv2
 import numpy as np
@@ -18,7 +18,7 @@ class CameraOpenCV(ICamera):
         self.camera_api = cv2.CAP_ANY
         self.camera = None
         self.camera_name = CAM_OPENCV
-        self.roi = [500, 500]
+        self.roi = CameraROI()
 
         # Windows platforms support discrete exposure times
         # These are mapped using a dictionary
@@ -38,6 +38,8 @@ class CameraOpenCV(ICamera):
             "244.1 us" : -12,
             "122.1 us" : -13
         }
+
+        self.frame_counter = 0
 
     def __del__(self) -> None:
         if self.camera is not None:
@@ -59,6 +61,7 @@ class CameraOpenCV(ICamera):
         _ , img = self.camera.read()
         img = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
         cv2.waitKey(1)
+        self.frame_counter += 1
         return img
 
     def set_exposure(self, exposure) -> None:
@@ -66,8 +69,26 @@ class CameraOpenCV(ICamera):
             exposure = self.exposure_dict[exposure]
         self.camera.set(cv2.CAP_PROP_EXPOSURE, exposure)
     
-    def set_roi(self, roi : list) -> None:
+    def set_roi(self, roi : CameraROI) -> None:
+        # todo: implement actual ROI
         self.roi = roi
     
-    def get_roi(self) -> list:
+    def get_roi(self) -> CameraROI:
         return self.roi
+    
+    def set_full_frame(self) -> None:
+        self.roi = CameraROI(0, 0, 500, 500)
+
+    def get_acquisition(self) -> bool:
+        return self.camera.isOpened()
+    
+    def set_acquisition(self, is_enabled) -> None:
+        if is_enabled:
+            self.camera = cv2.VideoCapture(self.camera_idx, self.camera_api)
+        else:
+            self.camera.release()
+    
+    def get_frames_per_second(self) -> int:
+        frames = self.frame_counter
+        self.frame_counter = 0
+        return frames