+++ napari_live_recording/Cameras/ICamera.py
@@ -2,7 +2,23 @@
 """
 
 from abc import ABC, abstractmethod
+from dataclasses import dataclass
 import numpy as np
+
+
+@dataclass
+class CameraROI():
+    """Dataclass for ROI settings.
+    """
+    offset_x: int = 0
+    offset_y: int = 0
+    height: int = 500
+    width: int = 500
+    ofs_x_step: int = 1
+    ofs_y_step: int = 1
+    width_step: int = 1
+    height_step: int = 1
+
 
 class ICamera(ABC):
     """Abstract class representing the camera device interface.
@@ -16,7 +32,6 @@ class ICamera(ABC):
     def __init__(self) -> None:
         """Abstract constructor method
         """
-        super().__init__()
         self.camera_name = "ICamera"
 
     @abstractmethod
@@ -32,7 +47,7 @@ class ICamera(ABC):
         :rtype: bool
         """
         pass
-    
+
     @abstractmethod
     def close_device(self) -> None:
         """Closes device, disabling acquisition
@@ -49,32 +64,98 @@ class ICamera(ABC):
         pass
 
     @abstractmethod
+    def get_available_pixel_formats(self) -> list:
+        """Returns a list with associated pixel byte type.
+
+        i.e.
+        ["8-bit gray", "16-bit gray"]
+
+        The first item of the list should be the default pixel format at camera startup.
+
+        :return: list of strings with the supported pixel formats.
+        :rtype: list[str]
+        """
+        pass
+
+    @abstractmethod
+    def set_pixel_format(self, format) -> None:
+        """Sets a new pixel data format for the camera.
+
+        :param format: the new data format (for OpenCV time scale is fixed, for other devices is microseconds)
+        :type exposure: str
+        """
+        pass
+
+    @abstractmethod
+    def get_pixel_format(self):
+        """Returns the current camera pixel format.
+        """
+        pass    
+
+    @abstractmethod
     def set_exposure(self, exposure) -> None:
         """Sets the exposure time for the current device.
 
         :param exposure: the new exposure time (for OpenCV time scale is fixed, for other devices is microseconds)
-        :type exposure: int, string
+        :type exposure: int, str
         """
         pass
 
     @abstractmethod
-    def set_roi(self, roi : list) -> None:
+    def set_roi(self, roi: CameraROI) -> None:
         """Sets a ROI for the current device.
 
-        :param roi: list of integers indicating the new ROI (this is still not supported)
-        :type roi: list
+        :param roi: tuple indicating the new ROI
+        :type roi: CameraROI
         """
         pass
 
     @abstractmethod
-    def get_roi(self) -> list:
+    def get_roi(self) -> CameraROI:
         """Returns the current set ROI.
 
-        :return: list of current ROI.
-        :rtype: list
+        :return: tuple of current ROI.
+        :rtype: CameraROI
         """
         pass
     
+    @abstractmethod
+    def get_sensor_range(self) -> CameraROI:
+        pass
+
+    @abstractmethod
+    def set_full_frame(self) -> None:
+        """Sets ROI to full frame.
+        """
+        pass
+
+    @abstractmethod
+    def get_acquisition(self) -> bool:
+        """Returns current acquisition status.
+
+        :return: True if camera is enabled, otherwise False.
+        :rtype: bool
+        """
+        pass
+
+    @abstractmethod
+    def set_acquisition(self, is_enabled) -> None:
+        """Sets current acquisition status.
+
+        :param is_enabled: new acquisition status
+        :type exposure: bool
+        """
+        pass
+
+    @abstractmethod
+    def get_frames_per_second(self) -> int:
+        """Returns the current FPS of a given device.
+
+        :return: the currently calculated FPS
+        :rtype: int
+        """
+        pass
+
     def get_name(self) -> str:
         """Returns the camera name specified by camera_name